Add ElasticCableBase with new implementation of assembly
We'll eventually reuse this function with Rayleigh damping, but for now,
make check is passing with the factored out function.
