+++ src/Makefile.am
@@ -123,6 +123,7 @@ libgrins_la_SOURCES += physics/src/elastic_membrane_abstract.C
 libgrins_la_SOURCES += physics/src/elastic_membrane.C
 libgrins_la_SOURCES += physics/src/elastic_membrane_constant_pressure.C
 libgrins_la_SOURCES += physics/src/elastic_cable_abstract.C
+libgrins_la_SOURCES += physics/src/elastic_cable_base.C
 libgrins_la_SOURCES += physics/src/elastic_cable.C
 libgrins_la_SOURCES += physics/src/elastic_cable_constant_gravity.C
 libgrins_la_SOURCES += physics/src/nonlinear_elasticity_instantiate.C
@@ -350,6 +351,7 @@ include_HEADERS += physics/include/grins/spalart_allmaras_spgsm_stab.h
 include_HEADERS += physics/include/grins/elastic_membrane_abstract.h
 include_HEADERS += physics/include/grins/elastic_membrane.h
 include_HEADERS += physics/include/grins/elastic_cable_abstract.h
+include_HEADERS += physics/include/grins/elastic_cable_base.h
 include_HEADERS += physics/include/grins/elastic_cable.h
 include_HEADERS += physics/include/grins/elastic_membrane_constant_pressure.h
 include_HEADERS += physics/include/grins/elastic_cable_constant_gravity.h
