+++ src/physics/include/grins/elastic_cable.h
@@ -27,12 +27,12 @@
 
 
 //GRINS
-#include "grins/elastic_cable_abstract.h"
+#include "grins/elastic_cable_base.h"
 
 namespace GRINS
 {
   template<typename StressStrainLaw>
-  class ElasticCable : public ElasticCableAbstract
+  class ElasticCable : public ElasticCableBase<StressStrainLaw>
   {
   public:
 
@@ -48,7 +48,11 @@ namespace GRINS
     //! Time dependent part(s) of physics for element interiors
     virtual void element_time_derivative( bool compute_jacobian,
                                           AssemblyContext& context,
-                                          CachedValues& cache );
+                                          CachedValues& /*cache*/ )
+    { this->element_time_derivative_impl(compute_jacobian,
+                                         context,
+                                         &libMesh::DiffContext::get_elem_solution,
+                                         &libMesh::DiffContext::get_elem_solution_derivative); }
 
     virtual void side_time_derivative( bool compute_jacobian,
 				       AssemblyContext& context,
@@ -67,22 +71,6 @@ namespace GRINS
   private:
 
     ElasticCable();
-
-    void compute_metric_tensors( unsigned int qp,
-                                 const libMesh::FEBase& elem,
-                                 const AssemblyContext& context,
-                                 const libMesh::Gradient& grad_u,
-                                 const libMesh::Gradient& grad_v,
-                                 const libMesh::Gradient& grad_w,
-                                 libMesh::TensorValue<libMesh::Real>& a_cov,
-                                 libMesh::TensorValue<libMesh::Real>& a_contra,
-                                 libMesh::TensorValue<libMesh::Real>& A_cov,
-                                 libMesh::TensorValue<libMesh::Real>& A_contra,
-                                 libMesh::Real& lambda_sq);
-
-    StressStrainLaw _stress_strain_law;
-
-    bool _is_compressible;
 
     //! Index from registering this quantity. Each component will have it's own index.
     std::vector<unsigned int> _stress_indices;
