+++ src/physics/src/elastic_cable.C
@@ -29,10 +29,7 @@
 #include "grins_config.h"
 #include "grins/solid_mechanics_bc_handling.h"
 #include "grins/generic_ic_handler.h"
-#include "grins/elasticity_tensor.h"
 #include "grins/postprocessed_quantities.h"
-#include "grins/physics.h"
-#include "grins/materials_parsing.h"
 
 // libMesh
 #include "libmesh/getpot.h"
@@ -46,15 +43,11 @@ namespace GRINS
   template<typename StressStrainLaw>
   ElasticCable<StressStrainLaw>::ElasticCable( const PhysicsName& physics_name, const GetPot& input,
                                                bool is_compressible )
-    : ElasticCableAbstract(physics_name,input),
-      _stress_strain_law(input,MaterialsParsing::material_name(input,PhysicsNaming::elastic_cable())),
-      _is_compressible(is_compressible)
+    : ElasticCableBase<StressStrainLaw>(physics_name,input,is_compressible)
   {
     this->_bc_handler = new SolidMechanicsBCHandling( physics_name, input );
 
     this->_ic_handler = new GenericICHandler(physics_name, input);
-
-    return;
   }
 
   template<typename StressStrainLaw>
@@ -113,152 +106,6 @@ namespace GRINS
   }
 
   template<typename StressStrainLaw>
-  void ElasticCable<StressStrainLaw>::element_time_derivative( bool compute_jacobian,
-                                                               AssemblyContext& context,
-                                                               CachedValues& /*cache*/ )
-  {
-    const unsigned int n_u_dofs = context.get_dof_indices(_disp_vars.u()).size();
-
-    const std::vector<libMesh::Real> &JxW =
-      this->get_fe(context)->get_JxW();
-
-    // Residuals that we're populating
-    libMesh::DenseSubVector<libMesh::Number> &Fu = context.get_elem_residual(_disp_vars.u());
-    libMesh::DenseSubVector<libMesh::Number> &Fv = context.get_elem_residual(_disp_vars.v());
-    libMesh::DenseSubVector<libMesh::Number> &Fw = context.get_elem_residual(_disp_vars.w());
-
-    //Grab the Jacobian matrix as submatrices
-    //libMesh::DenseMatrix<libMesh::Number> &K = context.get_elem_jacobian();
-    libMesh::DenseSubMatrix<libMesh::Number> &Kuu = context.get_elem_jacobian(_disp_vars.u(),_disp_vars.u());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kuv = context.get_elem_jacobian(_disp_vars.u(),_disp_vars.v());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kuw = context.get_elem_jacobian(_disp_vars.u(),_disp_vars.w());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kvu = context.get_elem_jacobian(_disp_vars.v(),_disp_vars.u());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kvv = context.get_elem_jacobian(_disp_vars.v(),_disp_vars.v());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kvw = context.get_elem_jacobian(_disp_vars.v(),_disp_vars.w());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kwu = context.get_elem_jacobian(_disp_vars.w(),_disp_vars.u());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kwv = context.get_elem_jacobian(_disp_vars.w(),_disp_vars.v());
-    libMesh::DenseSubMatrix<libMesh::Number> &Kww = context.get_elem_jacobian(_disp_vars.w(),_disp_vars.w());
-
-
-    unsigned int n_qpoints = context.get_element_qrule().n_points();
-
-    // All shape function gradients are w.r.t. master element coordinates
-    const std::vector<std::vector<libMesh::Real> >& dphi_dxi = this->get_fe(context)->get_dphidxi();
-
-    const libMesh::DenseSubVector<libMesh::Number>& u_coeffs = context.get_elem_solution( _disp_vars.u() );
-    const libMesh::DenseSubVector<libMesh::Number>& v_coeffs = context.get_elem_solution( _disp_vars.v() );
-    const libMesh::DenseSubVector<libMesh::Number>& w_coeffs = context.get_elem_solution( _disp_vars.w() );
-
-    // Need these to build up the covariant and contravariant metric tensors
-    const std::vector<libMesh::RealGradient>& dxdxi  = this->get_fe(context)->get_dxyzdxi();
-
-    const unsigned int dim = 1; // The cable dimension is always 1 for this physics
-
-    for (unsigned int qp=0; qp != n_qpoints; qp++)
-      {
-        // Gradients are w.r.t. master element coordinates
-        libMesh::Gradient grad_u, grad_v, grad_w;
-
-        for( unsigned int d = 0; d < n_u_dofs; d++ )
-          {
-            libMesh::RealGradient u_gradphi( dphi_dxi[d][qp] );
-            grad_u += u_coeffs(d)*u_gradphi;
-            grad_v += v_coeffs(d)*u_gradphi;
-            grad_w += w_coeffs(d)*u_gradphi;
-          }
-
-        libMesh::RealGradient grad_x( dxdxi[qp](0) );
-        libMesh::RealGradient grad_y( dxdxi[qp](1) );
-        libMesh::RealGradient grad_z( dxdxi[qp](2) );
-
-        libMesh::TensorValue<libMesh::Real> a_cov, a_contra, A_cov, A_contra;
-        libMesh::Real lambda_sq = 0;
-
-        this->compute_metric_tensors( qp, *(this->get_fe(context)), context,
-                                      grad_u, grad_v, grad_w,
-                                      a_cov, a_contra, A_cov, A_contra,
-                                      lambda_sq );
-
-        // Compute stress tensor
-        libMesh::TensorValue<libMesh::Real> tau;
-        ElasticityTensor C;
-        //_stress_strain_law.compute_stress(dim,a_contra,a_cov,A_contra,A_cov,tau);
-        _stress_strain_law.compute_stress_and_elasticity(dim,a_contra,a_cov,A_contra,A_cov,tau,C);
-
-
-        libMesh::Real jac = JxW[qp];
-
-        for (unsigned int i=0; i != n_u_dofs; i++)
-          {
-            libMesh::RealGradient u_gradphi( dphi_dxi[i][qp] );
-
-            Fu(i) += tau(0,0)*_A*( (grad_x(0) + grad_u(0))*u_gradphi(0) ) * jac;
-
-            Fv(i) += tau(0,0)*_A*( (grad_y(0) + grad_v(0))*u_gradphi(0) ) * jac;
-
-            Fw(i) += tau(0,0)*_A*( (grad_z(0) + grad_w(0))*u_gradphi(0) ) * jac;
-
-          }
-
-        if( compute_jacobian )
-          {
-            for(unsigned int i=0; i != n_u_dofs; i++)
-              {
-                libMesh::RealGradient u_gradphi_I( dphi_dxi[i][qp] );
-                for(unsigned int j=0; j != n_u_dofs; j++)
-                  {
-                    libMesh::RealGradient u_gradphi_J( dphi_dxi[j][qp] );
-
-                    const libMesh::Real diag_term = _A*jac*tau(0,0)*( u_gradphi_J(0)*u_gradphi_I(0));
-
-                    Kuu(i,j) += diag_term;
-
-                    Kvv(i,j) += diag_term;
-
-                    Kww(i,j) += diag_term;
-
-                    const libMesh::Real dgamma_du = ( u_gradphi_J(0)*(grad_x(0)+grad_u(0)) );
-
-                    const libMesh::Real dgamma_dv = ( u_gradphi_J(0)*(grad_y(0)+grad_v(0)) );
-
-                    const libMesh::Real dgamma_dw = ( u_gradphi_J(0)*(grad_z(0)+grad_w(0)) );
-
-                    const libMesh::Real C1 = _A*jac*C(0,0,0,0);
-
-                    const libMesh::Real x_term = C1*( (grad_x(0)+grad_u(0))*u_gradphi_I(0) );
-
-                    const libMesh::Real y_term = C1*( (grad_y(0)+grad_v(0))*u_gradphi_I(0) );
-
-                    const libMesh::Real z_term = C1*( (grad_z(0)+grad_w(0))*u_gradphi_I(0) );
-
-                    Kuu(i,j) += x_term*dgamma_du;
-
-                    Kuv(i,j) += x_term*dgamma_dv;
-
-                    Kuw(i,j) += x_term*dgamma_dw;
-
-                    Kvu(i,j) += y_term*dgamma_du;
-
-                    Kvv(i,j) += y_term*dgamma_dv;
-
-                    Kvw(i,j) += y_term*dgamma_dw;
-
-                    Kwu(i,j) += z_term*dgamma_du;
-
-                    Kwv(i,j) += z_term*dgamma_dv;
-
-                    Kww(i,j) += z_term*dgamma_dw;
-
-                  }
-              }
-            //libmesh_not_implemented();
-          }
-      }
-
-    return;
-  }
-
-  template<typename StressStrainLaw>
   void ElasticCable<StressStrainLaw>::side_time_derivative( bool compute_jacobian,
                                                             AssemblyContext& context,
                                                             CachedValues& cache )
@@ -280,7 +127,7 @@ namespace GRINS
                                                      AssemblyContext& context,
                                                      CachedValues& /*cache*/ )
   {
-    const unsigned int n_u_dofs = context.get_dof_indices(_disp_vars.u()).size();
+    const unsigned int n_u_dofs = context.get_dof_indices(this->_disp_vars.u()).size();
 
     const std::vector<libMesh::Real> &JxW =
       this->get_fe(context)->get_JxW();
@@ -289,13 +136,13 @@ namespace GRINS
       this->get_fe(context)->get_phi();
 
     // Residuals that we're populating
-    libMesh::DenseSubVector<libMesh::Number> &Fu = context.get_elem_residual(_disp_vars.u());
-    libMesh::DenseSubVector<libMesh::Number> &Fv = context.get_elem_residual(_disp_vars.v());
-    libMesh::DenseSubVector<libMesh::Number> &Fw = context.get_elem_residual(_disp_vars.w());
+    libMesh::DenseSubVector<libMesh::Number> &Fu = context.get_elem_residual(this->_disp_vars.u());
+    libMesh::DenseSubVector<libMesh::Number> &Fv = context.get_elem_residual(this->_disp_vars.v());
+    libMesh::DenseSubVector<libMesh::Number> &Fw = context.get_elem_residual(this->_disp_vars.w());
 
-    libMesh::DenseSubMatrix<libMesh::Number>& Kuu = context.get_elem_jacobian(_disp_vars.u(),_disp_vars.u());
-    libMesh::DenseSubMatrix<libMesh::Number>& Kvv = context.get_elem_jacobian(_disp_vars.v(),_disp_vars.v());
-    libMesh::DenseSubMatrix<libMesh::Number>& Kww = context.get_elem_jacobian(_disp_vars.w(),_disp_vars.w());
+    libMesh::DenseSubMatrix<libMesh::Number>& Kuu = context.get_elem_jacobian(this->_disp_vars.u(),this->_disp_vars.u());
+    libMesh::DenseSubMatrix<libMesh::Number>& Kvv = context.get_elem_jacobian(this->_disp_vars.v(),this->_disp_vars.v());
+    libMesh::DenseSubMatrix<libMesh::Number>& Kww = context.get_elem_jacobian(this->_disp_vars.w(),this->_disp_vars.w());
 
     unsigned int n_qpoints = context.get_element_qrule().n_points();
 
@@ -304,21 +151,21 @@ namespace GRINS
         libMesh::Real jac = JxW[qp];
 
         libMesh::Real u_ddot, v_ddot, w_ddot;
-        context.interior_accel( _disp_vars.u(), qp, u_ddot );
-        context.interior_accel( _disp_vars.v(), qp, v_ddot );
-        context.interior_accel( _disp_vars.w(), qp, w_ddot );
+        context.interior_accel( this->_disp_vars.u(), qp, u_ddot );
+        context.interior_accel( this->_disp_vars.v(), qp, v_ddot );
+        context.interior_accel( this->_disp_vars.w(), qp, w_ddot );
 
         for (unsigned int i=0; i != n_u_dofs; i++)
 	  {
-            Fu(i) += this->_rho*_A*u_ddot*u_phi[i][qp]*jac;
-            Fv(i) += this->_rho*_A*v_ddot*u_phi[i][qp]*jac;
-            Fw(i) += this->_rho*_A*w_ddot*u_phi[i][qp]*jac;
+            Fu(i) += this->_rho*this->_A*u_ddot*u_phi[i][qp]*jac;
+            Fv(i) += this->_rho*this->_A*v_ddot*u_phi[i][qp]*jac;
+            Fw(i) += this->_rho*this->_A*w_ddot*u_phi[i][qp]*jac;
 
             if( compute_jacobian )
               {
                 for (unsigned int j=0; j != n_u_dofs; j++)
                   {
-                    libMesh::Real jac_term = this->_rho*_A*u_phi[i][qp]*u_phi[j][qp]*jac;
+                    libMesh::Real jac_term = this->_rho*this->_A*u_phi[i][qp]*u_phi[j][qp]*jac;
                     jac_term *= context.get_elem_solution_accel_derivative();
 
                     Kuu(i,j) += jac_term;
@@ -350,11 +197,11 @@ namespace GRINS
 
     if( is_strain || is_stress || is_force )
       {
-        const unsigned int n_u_dofs = context.get_dof_indices(_disp_vars.u()).size();
+        const unsigned int n_u_dofs = context.get_dof_indices(this->_disp_vars.u()).size();
 
-        const libMesh::DenseSubVector<libMesh::Number>& u_coeffs = context.get_elem_solution( _disp_vars.u() );
-        const libMesh::DenseSubVector<libMesh::Number>& v_coeffs = context.get_elem_solution( _disp_vars.v() );
-        const libMesh::DenseSubVector<libMesh::Number>& w_coeffs = context.get_elem_solution( _disp_vars.w() );
+        const libMesh::DenseSubVector<libMesh::Number>& u_coeffs = context.get_elem_solution( this->_disp_vars.u() );
+        const libMesh::DenseSubVector<libMesh::Number>& v_coeffs = context.get_elem_solution( this->_disp_vars.v() );
+        const libMesh::DenseSubVector<libMesh::Number>& w_coeffs = context.get_elem_solution( this->_disp_vars.w() );
 
         // Build new FE for the current point. We need this to build tensors at point.
         libMesh::AutoPtr<libMesh::FEGenericBase<libMesh::Real> > fe_new =  this->build_new_fe( context.get_elem(), this->get_fe(context), point );
@@ -389,7 +236,7 @@ namespace GRINS
         libMesh::Real I3 = lambda_sq*det_A/det_a;
 
         libMesh::TensorValue<libMesh::Real> tau;
-        _stress_strain_law.compute_stress(1,a_contra,a_cov,A_contra,A_cov,tau);
+        this->_stress_strain_law.compute_stress(1,a_contra,a_cov,A_contra,A_cov,tau);
 
         // We have everything we need for strain now, so check if we are computing strain
         if( is_strain )
@@ -424,7 +271,7 @@ namespace GRINS
             if( _force_indices[0] == quantity_index )
               {
             	//This force is in deformed configuration and will be significantly influenced by large strains
-                value = tau(0,0)/std::sqrt(I3)*_A;
+                value = tau(0,0)/std::sqrt(I3)*this->_A;
               }
             else
               {
@@ -432,72 +279,6 @@ namespace GRINS
               }
           }
       }
-
-    return;
-  }
-
-  template<typename StressStrainLaw>
-  void ElasticCable<StressStrainLaw>::compute_metric_tensors( unsigned int qp,
-                                                              const libMesh::FEBase& elem,
-                                                              const AssemblyContext& /*context*/,
-                                                              const libMesh::Gradient& grad_u,
-                                                              const libMesh::Gradient& grad_v,
-                                                              const libMesh::Gradient& grad_w,
-                                                              libMesh::TensorValue<libMesh::Real>& a_cov,
-                                                              libMesh::TensorValue<libMesh::Real>& a_contra,
-                                                              libMesh::TensorValue<libMesh::Real>& A_cov,
-                                                              libMesh::TensorValue<libMesh::Real>& A_contra,
-                                                              libMesh::Real& lambda_sq )
-  {
-    const std::vector<libMesh::RealGradient>& dxdxi  = elem.get_dxyzdxi();
-
-    const std::vector<libMesh::Real>& dxidx  = elem.get_dxidx();
-    const std::vector<libMesh::Real>& dxidy  = elem.get_dxidy();
-    const std::vector<libMesh::Real>& dxidz  = elem.get_dxidz();
-
-    libMesh::RealGradient dxi( dxidx[qp], dxidy[qp], dxidz[qp] );
-
-    libMesh::RealGradient dudxi( grad_u(0), grad_v(0), grad_w(0) );
-
-    // Covariant metric tensor of reference configuration
-    a_cov.zero();
-    a_cov(0,0) = dxdxi[qp]*dxdxi[qp];
-    a_cov(1,1)    = 1.0;
-    a_cov(2,2)    = 1.0;
-
-    // Covariant metric tensor of current configuration
-    A_cov.zero();
-    A_cov(0,0) = (dxdxi[qp] + dudxi)*(dxdxi[qp] + dudxi);
-
-    // Contravariant metric tensor of reference configuration
-    a_contra.zero();
-    a_contra(0,0) = 1/a_cov(0,0);
-    a_contra(1,1) = 1.0;
-    a_contra(2,2) = 1.0;
-
-    // Contravariant metric tensor in current configuration is A_cov^{-1}
-    A_contra.zero();
-    A_contra(0,0) =  1/A_cov(0,0);
-
-    // If the material is compressible, then lambda_sq is an independent variable
-    if( _is_compressible )
-      {
-        libmesh_not_implemented();
-        //lambda_sq = context.interior_value(this->_lambda_sq_var, qp);
-      }
-    else
-      {
-        // If the material is incompressible, lambda^2 is known
-        lambda_sq = a_cov(0,0)/A_cov(0,0);//det_a/det_A;
-      }
-
-    //Update the covariant and contravariant tensors of current configuration
-    A_cov(1,1)    = lambda_sq;
-    A_cov(2,2)    = lambda_sq;
-    A_contra(1,1) = 1.0/lambda_sq;
-    A_contra(2,2) = 1.0/lambda_sq;
-
-    return;
   }
 
 } // end namespace GRINS
